requirements.txt
pandas
matplotlib
pillow
fastapi
uvicorn[standard]
fhirclient
python-dotenv
pytest
